+++ pyClarion/components/chunks_.py
@@ -21,9 +21,10 @@ from typing import (
     List, Optional, TypeVar, Type, Generic, MutableMapping, cast, overload, 
     Any, Dict
 )
+from contextlib import contextmanager
 from collections import namedtuple
 from statistics import mean
-from itertools import count
+from itertools import count, chain
 from copy import copy
 from dataclasses import dataclass
 from types import MappingProxyType
@@ -126,6 +127,18 @@ class Chunk(object):
 
         return strength
 
+    def support(self, *domains: Domain) -> bool:
+        """
+        Return True iff domains support self.
+        
+        A set of domains is considered to support a chunk if every feature of 
+        the chunk is in at least one of the domains.
+        """
+        
+        vals = all(any(f in d.features for d in domains) for f in self.features)   
+        
+        return vals
+
 
 Ct = TypeVar("Ct", bound="Chunk")
 class Chunks(MutableMapping[chunk, Ct]):
@@ -164,6 +177,8 @@ class Chunks(MutableMapping[chunk, Ct]):
         else:
             _data = dict(data)
         self._data: MutableMapping[chunk, Ct] = _data
+        self._domains: Tuple[Domain, ...] = ()
+        self._enforce_support: bool = False
 
         if chunk_type is None:
             self.Chunk = Chunk
@@ -197,6 +212,9 @@ class Chunks(MutableMapping[chunk, Ct]):
     def __setitem__(self, key, val):
 
         if isinstance(val, self.Chunk):
+            if self._enforce_support and not val.support(*self._domains):
+                msg = "Chunk {} contains unexpected features."
+                raise ValueError(msg.format(key.cid))
             self._data[key] = val
         else:
             msg = "This chunk database expects chunks of type '{}'." 
@@ -256,18 +274,20 @@ class Chunks(MutableMapping[chunk, Ct]):
 
         return 0 < len(found) 
 
-    def support(self, *domains: Domain) -> bool:
+    @contextmanager
+    def enforce_support(self, *domains: Domain):
         """
-        Return True iff domains support all chunks in self.
+        Reject new chunks that are not supported by domains.
         
-        A set of domains is considered to support a chunk if every feature of 
-        the chunk is in at least one of the domains.
+        Use this method as a context manager when initializing complex 
+        knowledge to automatically check for coding errors.
         """
 
-        s = {f for dom in domains for f in dom.features}
-        value = all(f in s for c in self.values() for f in c.features)
-
-        return value
+        self._domains = domains
+        self._enforce_support = True
+        yield
+        self._enforce_support = False
+        self._domains = ()
 
     def request_add(self, ch, form):
         """